+++ Source/Bake/Services/DestinationParser.cs
@@ -22,6 +22,7 @@
 
 using System.Text.RegularExpressions;
 using Bake.Core;
+using Bake.ValueObjects.Artifacts;
 using Bake.ValueObjects.Destinations;
 
 namespace Bake.Services
@@ -59,8 +60,7 @@ namespace Bake.Services
                             Names.DynamicDestinations.GitHub);
                         return true;
                     }
-
-                    return false;
+                    break;
 
                 case Names.ArtifactTypes.NuGet:
                     if (parts.Length == 1)
@@ -130,9 +130,16 @@ namespace Bake.Services
                     destination = new ContainerRegistryDestination($"{parts[1].TrimEnd('/')}/");
                     return true;
 
-                default:
-                    return false;
+                case Names.ArtifactTypes.DocumentationSite:
+                    if (string.Equals(parts[1], Names.Destinations.Container))
+                    {
+                        destination = new ContainerDestination(ArtifactType.DocumentationSite);
+                        return true;
+                    }
+                    break;
             }
+
+            return false;
         }
     }
 }