+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # 0.23-beta
 
+* Feature: Output a container for hosting mkdocs documentation
 * Feature: Add standard labels from the Open Containers Initiative
 * Feature: Add dumb-init to .NET containers to ensure that the process
   and all child-processes are correctly terminated